+++ webpages/settings/components/addon-body.html
@@ -1,0 +1,80 @@
+<template>
+  <div class="addon-body" v-show="shouldShow && groupExpanded" :id="'addon-' + addon._addonId">
+    <div class="addon-topbar">
+      <div class="clickeable-area" @click="expanded = !expanded">
+        <div class="btn-dropdown">
+          <img src="../../images/icons/expand.svg" alt="v" :class="{'reverted': expanded}" />
+        </div>
+        <div class="addon-name">
+          <img v-if="addon._icon === 'editor'" src="../../images/icons/puzzle.svg" class="icon-type" />
+          <img v-if="addon._icon === 'community'" src="../../images/icons/web.svg" class="icon-type" />
+          <img v-if="addon._icon === 'theme'" src="../../images/icons/brush.svg" class="icon-type" />
+          <img v-if="addon._icon === 'easterEgg'" src="../../images/icons/egg-easter.svg" class="icon-type" />
+          <img v-if="addon._icon === 'popup'" src="../../images/icons/popup.svg" class="icon-type" />
+          <span @click="devShowAddonIds">{{ addon._displayedAddonId || addon.name }}</span>
+        </div>
+      </div>
+      <addon-tag v-for="tag of addon.tags" :tag="tag"></addon-tag>
+      <div class="addon-description" v-show="!expanded" dir="auto">{{ addon.description }}</div>
+      <div class="addon-check">
+        <div
+          v-show="expanded && addon._enabled"
+          v-if="addon.settings"
+          class="clear-settings"
+          title="{{ msg('resetToDefault') }}"
+          @click="loadDefaults"
+        >
+          <img src="../../images/icons/undo.svg" class="icon-type" />
+        </div>
+        <div class="switch" :state="addon._enabled ? 'on' : 'off'" @click="toggleAddonRequest($event)"></div>
+      </div>
+    </div>
+    <div class="addon-settings" v-if="expanded">
+      <div class="addon-description-full">{{ addon.description }}</div>
+      <div id="info" v-for="info of addon.info">
+        <div class="addon-warn" v-if="info.type == 'warning'">
+          <img src="../../images/icons/warning.svg" />{{ info.text }}
+        </div>
+        <div class="addon-notice" v-if="info.type == 'notice'">
+          <img src="../../images/icons/notice.svg" />{{ info.text }}
+        </div>
+      </div>
+      <div class="addon-credits" v-if="addon.credits">
+        <span>{{ msg("creditTo") }}</span>
+        <span v-for="credit of addon.credits">
+          <span v-if="credit.link">
+            <a href="{{ credit.link }}" rel="noreferrer noopener" target="_blank">{{ credit.name }}</a>
+          </span>
+          <span v-else="credit.link">{{ credit.name }}</span>
+        </span>
+      </div>
+      <div class="addon-license" v-if="addon.libraries && addon.libraries.length">
+        <a target="_blank" href="./licenses.html?libraries={{ addon.libraries.join(',') }}"
+          >{{ msg("viewLicenses") }}</a
+        >
+      </div>
+      <div class="settings-column" v-bind:class="[!addon._enabled ? 'disabled' : '']">
+        <addon-setting
+          v-for="setting of addon.settings"
+          :addon="addon"
+          :setting="setting"
+          :addon-settings="addonSettings"
+        ></addon-setting>
+      </div>
+      <div class="presets-column" v-if="addon.presets" v-bind:class="[!addon._enabled ? 'disabled' : '']">
+        <div class="setting-label">{{ msg("presets") }}</div>
+        <div class="addon-setting" v-for="preset of addon.presets">
+          <button
+            type="preset-button"
+            class="large-button"
+            :disabled="!addon._enabled"
+            @click="loadPreset(preset)"
+            :title="preset.description"
+          >
+            {{preset.name}}
+          </button>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
