+++ webpages/settings/components/addon-group.html
@@ -1,0 +1,14 @@
+<template>
+  <div class="addon-group" v-show="shouldShow">
+    <span @click="toggle">
+      <img src="../../images/icons/expand.svg" alt="v" :class="{ 'reverted': group.expanded }" />
+      {{ group.name }} ({{ shownCount }})</span
+    >
+    <addon-body
+      v-for="addonId of group.addonIds"
+      :addon="manifestsById[addonId]"
+      :group-id="group.id"
+      :group-expanded="group.expanded"
+    ></addon-body>
+  </div>
+</template>
