+++ webpages/settings/components/addon-setting.html
@@ -1,0 +1,98 @@
+<template>
+  <div class="addon-setting">
+    <div class="setting-label" v-html="settingsName(addon)"></div>
+    <input
+      type="checkbox"
+      class="setting-input check"
+      v-model="addonSettings[addon._addonId][setting.id]"
+      @change="updateSettings()"
+      :disabled="!addon._enabled"
+      v-if="setting.type === 'boolean'"
+    />
+    <input
+      type="number"
+      class="setting-input number"
+      v-model="addonSettings[addon._addonId][setting.id]"
+      @change="checkValidity() || updateSettings()"
+      :disabled="!addon._enabled"
+      v-if="setting.type === 'positive_integer'"
+      min="0"
+      :data-addon-id="addon._addonId"
+      :data-setting-id="setting.id"
+      number
+    />
+    <input
+      type="number"
+      class="setting-input number"
+      v-model="addonSettings[addon._addonId][setting.id]"
+      @change="checkValidity() || updateSettings()"
+      :disabled="!addon._enabled"
+      v-if="setting.type === 'integer'"
+      :min="setting.min"
+      :max="setting.max"
+      :data-addon-id="addon._addonId"
+      :data-setting-id="setting.id"
+      number
+    />
+    <input
+      type="text"
+      class="setting-input string"
+      v-model="addonSettings[addon._addonId][setting.id]"
+      @input="updateSettings()"
+      :disabled="!addon._enabled"
+      v-if="setting.type === 'string'"
+      :placeholder="setting.default"
+      maxlength="100"
+    />
+    <input
+      type="text"
+      class="setting-input key"
+      v-model="addonSettings[addon._addonId][setting.id]"
+      @input="updateSettings()"
+      :disabled="!addon._enabled"
+      v-if="setting.type === 'key'"
+      :placeholder="setting.default"
+      maxlength="100"
+      :data-addon-id="addon._addonId"
+      :data-setting-id="setting.id"
+      spellcheck="false"
+    />
+    <template v-if="setting.type === 'color'">
+      <picker
+        :value="addonSettings[addon._addonId][setting.id] || setting.default"
+        :setting="setting"
+        :addon="addon"
+        :no_alpha="!setting.allowTransparency"
+        v-click-outside="closePickers"
+      ></picker>
+      <reset-dropdown
+        v-if="showResetDropdown()"
+        :setting="setting"
+        :addon="addon"
+        :label="msg('reset')"
+        :default-label="msg('default')"
+        v-click-outside="closeResetDropdowns"
+      ></reset-dropdown>
+      <button
+        v-else
+        type="button"
+        class="large-button clear-button"
+        :disabled="!addon._enabled"
+        @click="updateOption(setting.default)"
+      >
+        <span>{{ msg("reset") }}</span>
+        <img src="../../images/icons/undo.svg" class="icon-type" />
+      </button>
+    </template>
+    <div class="filter-options" v-if="setting.type === 'select'">
+      <div
+        class="filter-option"
+        v-for="option of setting.potentialValues"
+        :class="{'sel' : addonSettings[addon._addonId][setting.id] === option.id, 'disabled': !addon._enabled}"
+        @click="updateOption(option.id);"
+      >
+        {{ option.name }}
+      </div>
+    </div>
+  </div>
+</template>
