+++ webpages/settings/components/picker-component.html
@@ -1,0 +1,22 @@
+<template id="picker-component">
+  <div class="color-container" :key="value" @click="loadColorPicker = true" @mouseover="loadColorPicker = true">
+    <button
+      :style="{'background-color': color }"
+      class="setting-input color"
+      :class="{'action-disabled': !addon._enabled }"
+      @click="toggle(addon, setting, true)"
+    >
+      <template v-if="loadColorPicker">
+        <color-picker
+          :value="value"
+          id="picker"
+          v-show="isOpen"
+          v-el:pickr
+          :formats="formats"
+          :no_alpha="noAlphaString"
+          dir="ltr"
+        ></color-picker>
+      </template>
+    </button>
+  </div>
+</template>
