+++ webpages/settings/components/reset-dropdown.html
@@ -1,0 +1,23 @@
+<template>
+  <div :class="{'setting-dropdown': true, open: isOpen}">
+    <button type="button" class="large-button clear-button" :disabled="!addon._enabled" @click="toggle">
+      <span>{{ label }}</span>
+      <img src="../../images/icons/expand.svg" class="icon-type" />
+    </button>
+    <ul>
+      <li @click="resetToDefault">
+        <span class="reset-preview"><span :style="{backgroundColor: setting.default}"></span></span>
+        <span>{{ defaultLabel }}</span>
+      </li>
+      <template v-for="preset in addon.presets">
+        <li
+          v-if="preset.values.hasOwnProperty(setting.id) && preset.values[setting.id].toLowerCase() != setting.default.toLowerCase()"
+          @click="resetToPreset(preset)"
+        >
+          <span class="reset-preview"><span :style="{backgroundColor: preset.values[setting.id]}"></span></span>
+          <span>{{ preset.name }}</span>
+        </li>
+      </template>
+    </ul>
+  </div>
+</template>
