+++ webpages/settings/index.js
@@ -1,5 +1,7 @@
 import downloadBlob from "../../libraries/common/cs/download-blob.js";
 import getDirection from "../rtl-list.js";
+import loadVueComponent from "../../libraries/common/load-vue-components.js";
+import tags from "./tags.js";
 
 let isIframe = false;
 if (window.parent !== window) {
@@ -7,6 +9,18 @@ if (window.parent !== window) {
   document.body.classList.add("iframe");
   isIframe = true;
 }
+
+(async () => {
+
+await loadVueComponent([
+    "webpages/settings/components/picker-component",
+    "webpages/settings/components/reset-dropdown",
+    "webpages/settings/components/addon-setting",
+    "webpages/settings/components/addon-tag",
+    "webpages/settings/components/addon-group",
+    "webpages/settings/components/addon-body",
+    "webpages/settings/components/category-selector",
+]);  
 
 Vue.directive("click-outside", {
   priority: 700,
@@ -27,438 +41,6 @@ Vue.directive("click-outside", {
     event.stopPropagation();
   },
 });
-
-const ColorInput = Vue.extend({
-  props: ["value", "addon", "setting", "no_alpha"],
-  template: document.querySelector("template#picker-component").innerHTML,
-  data() {
-    return {
-      isOpen: false,
-      color: this.value,
-      canCloseOutside: false,
-      formats: "",
-      opening: false,
-      loadColorPicker: false, // #2090 tempfix
-    };
-  },
-  ready() {
-    if (!this.loadColorPicker) return;
-    if (this.no_alpha) {
-      this.formats = "hex,rgb,hsv,hsl";
-    } else {
-      this.formats = "hex,hex8,rgb,hsv,hsl";
-    }
-    this.$els.pickr.addEventListener("input", (e) => {
-      this.color = "#" + e.detail.value;
-      if (this.value !== this.color) {
-        this.$parent.addonSettings[this.addon._addonId][this.setting.id] = "#" + this.$els.pickr.hex8;
-        this.$parent.updateSettings(this.addon, { wait: 250, settingId: this.setting.id });
-      }
-    });
-    this.$root.$on("close-pickers", (except) => {
-      if (this.isOpen && this !== except) {
-        const addon = this.$parent.addon;
-        const setting = this.$parent.setting;
-        this.toggle(addon, setting, false, {
-          // I trust callers
-          callCloseDropdowns: false,
-          callClosePickers: false,
-        });
-      }
-    });
-  },
-  computed: {
-    noAlphaString() {
-      return String(this.no_alpha);
-    },
-  },
-  methods: {
-    toggle(addon, setting, value = !this.isOpen, { callCloseDropdowns = true, callClosePickers = true } = {}) {
-      if (!this.loadColorPicker) return;
-      this.isOpen = value;
-      this.opening = true;
-      if (callClosePickers)
-        this.$root.closePickers({ isTrusted: true }, this, {
-          callCloseDropdowns: false,
-        });
-      if (callCloseDropdowns) this.$root.closeResetDropdowns({ isTrusted: true }); // close other dropdowns
-      this.opening = false;
-      this.$els.pickr._valueChanged();
-      this.color = "#" + this.$els.pickr.hex8;
-      if (this.value !== this.color) {
-        this.$parent.addonSettings[addon._addonId][setting.id] = "#" + this.$els.pickr.hex8;
-        this.$parent.updateSettings(addon, { wait: 250, settingId: setting.id });
-      }
-      this.canCloseOutside = false;
-      setTimeout(() => {
-        this.canCloseOutside = true;
-      }, 0);
-    },
-  },
-  watch: {
-    value() {
-      this.color = this.value;
-      // ?. is #2090 tempfix, 4 lines below as well
-      this.$els.pickr?._valueChanged();
-    },
-    isOpen() {
-      this.$els.pickr?._valueChanged();
-    },
-    loadColorPicker() {
-      this.$options.ready[0].call(this);
-    },
-  },
-});
-Vue.component("picker", ColorInput);
-
-const ResetDropdown = Vue.extend({
-  props: ["addon", "setting", "label", "defaultLabel"],
-  template: document.querySelector("template#reset-dropdown-component").innerHTML,
-  data() {
-    return {
-      isResetDropdown: true,
-      isOpen: false,
-    };
-  },
-  ready() {
-    this.$root.$on("close-reset-dropdowns", (except) => {
-      if (this.isOpen && this !== except) {
-        this.isOpen = false;
-      }
-    });
-  },
-  methods: {
-    toggle() {
-      this.isOpen = !this.isOpen;
-      this.$root.closePickers({ isTrusted: true }, null, {
-        callCloseDropdowns: false,
-      });
-      this.$root.closeResetDropdowns({ isTrusted: true }, this); // close other dropdowns
-    },
-    resetToDefault() {
-      this.$parent.addonSettings[this.addon._addonId][this.setting.id] = this.setting.default;
-      this.$parent.updateSettings(this.addon, { settingId: this.setting.id });
-      this.toggle();
-    },
-    resetToPreset(preset) {
-      this.$parent.addonSettings[this.addon._addonId][this.setting.id] = preset.values[this.setting.id];
-      this.$parent.updateSettings(this.addon, { settingId: this.setting.id });
-      this.toggle();
-    },
-  },
-});
-Vue.component("reset-dropdown", ResetDropdown);
-
-const AddonGroup = Vue.extend({
-  props: ["group"],
-  template: document.querySelector("template#addon-group-component").innerHTML,
-  data() {
-    return {};
-  },
-  computed: {
-    shouldShow() {
-      if (this.group.id === "new" && this.$root.searchInput !== "") return false;
-      return this.shownCount > 0;
-    },
-    shownCount() {
-      // Recompute after root loaded
-      void this.$root.loaded;
-
-      return this.$children.filter((addon) => addon.shouldShow).length;
-    },
-    manifestsById() {
-      return this.$root.manifestsById;
-    },
-  },
-  methods: {
-    toggle() {
-      this.group.expanded = !this.group.expanded;
-    },
-  },
-});
-Vue.component("addon-group", AddonGroup);
-
-const AddonBody = Vue.extend({
-  props: ["addon", "groupId", "groupExpanded"],
-  template: document.querySelector("template#addon-body-component").innerHTML,
-  data() {
-    return {
-      expanded: isIframe ? false : this.groupId === "enabled",
-    };
-  },
-  computed: {
-    shouldShow() {
-      const matches =
-        this.$root.selectedCategory === "all" || this.addon._categories.includes(this.$root.selectedCategory);
-      if (matches) return this.addonMatchesFilters;
-      else return false;
-    },
-    searchInput() {
-      return this.$root.searchInput;
-    },
-    addonSettings() {
-      return this.$root.addonSettings;
-    },
-    addonMatchesFilters() {
-      if (this.groupId !== "search" && this.searchInput !== "") return false;
-      if (!this.addon._wasEverEnabled) this.addon._wasEverEnabled = this.addon._enabled;
-
-      const search = this.searchInput.toLowerCase();
-
-      const matchesSearch =
-        this.searchInput === "" ||
-        this.addon.name.toLowerCase().includes(search) ||
-        this.addon._addonId.toLowerCase().includes(search) ||
-        this.addon.description.toLowerCase().includes(search) ||
-        (this.addon.credits &&
-          this.addon.credits.map((obj) => obj.name.toLowerCase()).some((author) => author.includes(search)));
-      // Show disabled easter egg addons only if category is easterEgg
-      const matchesEasterEgg = this.addon.tags.includes("easterEgg")
-        ? this.$root.selectedCategory === "easterEgg" || this.addon._wasEverEnabled
-        : true;
-
-      return matchesSearch && matchesEasterEgg;
-    },
-  },
-  methods: {
-    devShowAddonIds(event) {
-      if (!this.$root.versionName.endsWith("-prerelease") || !event.ctrlKey) return;
-      event.stopPropagation();
-      Vue.set(this.addon, "_displayedAddonId", this.addon._addonId);
-    },
-    loadPreset(preset) {
-      if (window.confirm(chrome.i18n.getMessage("confirmPreset"))) {
-        for (const property of Object.keys(preset.values)) {
-          this.$root.addonSettings[this.addon._addonId][property] = preset.values[property];
-        }
-        this.$root.updateSettings(this.addon);
-        console.log(`Loaded preset ${preset.id} for ${this.addon._addonId}`);
-      }
-    },
-    loadDefaults() {
-      if (window.confirm(chrome.i18n.getMessage("confirmReset"))) {
-        for (const property of this.addon.settings) {
-          this.$root.addonSettings[this.addon._addonId][property.id] = property.default;
-        }
-        this.$root.updateSettings(this.addon);
-        console.log(`Loaded default values for ${this.addon._addonId}`);
-      }
-    },
-    toggleAddonRequest(event) {
-      const toggle = () => {
-        // Prevents selecting text when the shift key is being held down
-        event.preventDefault();
-
-        const newState = !this.addon._enabled;
-        this.addon._enabled = newState;
-        // Do not extend when enabling in popup mode, unless addon has warnings
-        this.expanded =
-          isIframe && !this.expanded && (this.addon.info || []).every((item) => item.type !== "warning")
-            ? false
-            : event.shiftKey
-            ? false
-            : newState;
-        chrome.runtime.sendMessage({ changeEnabledState: { addonId: this.addon._addonId, newState } });
-      };
-
-      const requiredPermissions = (this.addon.permissions || []).filter((value) =>
-        browserLevelPermissions.includes(value)
-      );
-      if (!this.addon._enabled && this.addon.tags.includes("danger")) {
-        const confirmation = confirm(chrome.i18n.getMessage("dangerWarning", [this.addon.name]));
-        if (!confirmation) return;
-      }
-      if (!this.addon._enabled && requiredPermissions.length) {
-        const result = requiredPermissions.every((p) => grantedOptionalPermissions.includes(p));
-        if (result === false) {
-          if (isIframe) {
-            this.$root.addonToEnable = this.addon;
-            document.querySelector(".popup").style.animation = "dropDown 1.6s 1";
-            this.$root.showPopupModal = true;
-          } else
-            chrome.permissions.request(
-              {
-                permissions: requiredPermissions,
-              },
-              (granted) => {
-                if (granted) {
-                  console.log("Permissions granted!");
-                  toggle();
-                }
-              }
-            );
-        } else toggle();
-      } else toggle();
-    },
-    msg(...params) {
-      return this.$root.msg(...params);
-    },
-  },
-});
-Vue.component("addon-body", AddonBody);
-
-// Tags in order
-const tags = [
-  {
-    name: "danger",
-    tooltipText: "dangerTooltip",
-    matchName: "danger",
-    color: "darkred",
-    iframeAlwaysShow: true,
-  },
-  {
-    name: "recommended",
-    tooltipText: "recommendedTooltip",
-    matchName: "recommended",
-    color: "blue",
-  },
-  {
-    name: "new",
-    matchName: "new",
-    color: "purple",
-  },
-  {
-    name: "beta",
-    tooltipText: "betaTooltip",
-    matchName: "beta",
-    color: "red",
-    iframeAlwaysShow: true,
-  },
-  {
-    name: "forums",
-    tooltipText: "forumsTooltip",
-    matchName: "forums",
-    color: "green",
-  },
-  {
-    name: "forEditor",
-    matchName: "editor",
-    color: "darkgreen",
-    addonTabShow: {
-      theme: true,
-    },
-  },
-  {
-    name: "forWebsite",
-    matchName: "community",
-    color: "yellow",
-    addonTabShow: {
-      theme: true,
-    },
-  },
-];
-const AddonTag = Vue.extend({
-  props: ["tag"],
-  template: document.querySelector("template#addon-tag-component").innerHTML,
-  data() {
-    return {};
-  },
-  computed: {
-    tagInfo() {
-      // Might return undefined, tag might not exist
-      return tags.find((tag) => tag.matchName === this.tag);
-    },
-    shouldShow() {
-      if (isIframe) return this.tagInfo && this.tagInfo.iframeAlwaysShow;
-      return this.tagInfo && (!this.tagInfo.addonTabShow || this.tagInfo.addonTabShow[this.$root.selectedCategory]);
-    },
-    tagName() {
-      return chrome.i18n.getMessage(this.tagInfo.name);
-    },
-    tagTooltip() {
-      return chrome.i18n.getMessage(this.tagInfo.tooltipText);
-    },
-  },
-});
-Vue.component("addon-tag", AddonTag);
-
-const AddonSetting = Vue.extend({
-  props: ["addon", "setting", "addon-settings"],
-  template: document.querySelector("template#addon-setting-component").innerHTML,
-  data() {
-    return {};
-  },
-  methods: {
-    settingsName(addon) {
-      const name = this.setting.name;
-      const regex = /([\\]*)(@|#)([a-zA-Z0-9.\-\/_]*)/g;
-      return name.replace(regex, (icon) => {
-        if (icon[0] === "\\") {
-          return icon.slice(1);
-        }
-        if (icon[0] === "@") {
-          return `<img class="inline-icon" src="../../images/icons/${icon.split("@")[1]}"/>`;
-        }
-        if (icon[0] === "#") {
-          return `<img class="inline-icon" src="../../addons/${addon._addonId}/${icon.split("#")[1]}"/>`;
-        }
-      });
-    },
-
-    showResetDropdown() {
-      return (
-        this.addon.presets &&
-        this.addon.presets.some(
-          (preset) => this.setting.id in preset.values && preset.values[this.setting.id] !== this.setting.default
-        )
-      );
-    },
-    checkValidity() {
-      // Needed to get just changed input to enforce it's min, max, and integer rule if the user "manually" sets the input to a value.
-      let input = document.querySelector(
-        `input[type='number'][data-addon-id='${this.addon._addonId}'][data-setting-id='${this.setting.id}']`
-      );
-      this.addonSettings[this.addon._addonId][this.setting.id] = input.validity.valid
-        ? input.value
-        : this.setting.default;
-    },
-    msg(...params) {
-      return this.$root.msg(...params);
-    },
-    updateSettings(...params) {
-      if (!params[0]) params[0] = this.addon;
-      this.$root.updateSettings(...params);
-    },
-    updateOption(newValue) {
-      this.$root.updateOption(this.setting.id, newValue, this.addon);
-    },
-  },
-  events: {
-    closePickers(...params) {
-      return this.$root.closePickers(...params);
-    },
-    closeResetDropdowns(...params) {
-      return this.$root.closeResetDropdowns(...params);
-    },
-  },
-});
-Vue.component("addon-setting", AddonSetting);
-
-const CategorySelector = Vue.extend({
-  props: ["category"],
-  template: document.querySelector("template#category-selector-component").innerHTML,
-  data() {
-    return {};
-  },
-  computed: {
-    selectedCategory() {
-      return this.$root.selectedCategory;
-    },
-    shouldShow() {
-      const categoriesWithParent = this.$root.categories
-        .filter((category) => category.parent === this.category.parent)
-        .map((category) => category.id);
-      return !this.category.parent || [this.category.parent, ...categoriesWithParent].includes(this.selectedCategory);
-    },
-  },
-  methods: {
-    onClick(event) {
-      event.stopPropagation();
-      this.$root.selectedCategory = this.category.id;
-    },
-  },
-});
-Vue.component("category-selector", CategorySelector);
 
 const browserLevelPermissions = ["notifications"];
 if (typeof browser !== "undefined") browserLevelPermissions.push("clipboardWrite");
@@ -1148,3 +730,5 @@ document.addEventListener("keydown", (e) => {
 });
 
 chrome.runtime.sendMessage("checkPermissions");
+
+})();
