+++ webpages/settings/tags.js
@@ -1,0 +1,50 @@
+// Tags in order
+export default [
+  {
+    name: "danger",
+    tooltipText: "dangerTooltip",
+    matchName: "danger",
+    color: "darkred",
+    iframeAlwaysShow: true,
+  },
+  {
+    name: "recommended",
+    tooltipText: "recommendedTooltip",
+    matchName: "recommended",
+    color: "blue",
+  },
+  {
+    name: "new",
+    matchName: "new",
+    color: "purple",
+  },
+  {
+    name: "beta",
+    tooltipText: "betaTooltip",
+    matchName: "beta",
+    color: "red",
+    iframeAlwaysShow: true,
+  },
+  {
+    name: "forums",
+    tooltipText: "forumsTooltip",
+    matchName: "forums",
+    color: "green",
+  },
+  {
+    name: "forEditor",
+    matchName: "editor",
+    color: "darkgreen",
+    addonTabShow: {
+      theme: true,
+    },
+  },
+  {
+    name: "forWebsite",
+    matchName: "community",
+    color: "yellow",
+    addonTabShow: {
+      theme: true,
+    },
+  },
+];
